extracted the frame handler from the relay socket
