+++ lib/nostr/relay_socket/frame_handler.ex
@@ -1,0 +1,29 @@
+defmodule Nostr.RelaySocket.FrameHandler do
+  require Logger
+
+  def handle_text_frame(frame, subscriptions) do
+    with {:ok, data} <- Jason.decode(frame),
+         {:ok, item} <- Nostr.Client.FrameDispatcher.dispatch(data) do
+      case get_atom_id(item) do
+        nil ->
+          :ok
+
+        atom_id ->
+          subscriber = Keyword.get(subscriptions, atom_id)
+          {_id, event} = item
+
+          send(subscriber, event)
+      end
+    else
+      {:error, _} -> Logger.warning("cannot parse frame: #{frame}")
+    end
+  end
+
+  defp get_atom_id({id, _}) do
+    String.to_atom(id)
+  end
+
+  defp get_atom_id(_) do
+    nil
+  end
+end
